pyproject.toml
[tool.poetry]
name = "morph-sample-project"
version = "0.1.0"
description = ""
authors = ["shibatanaoto <naoto.shibata510@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
morph-data = "0.3.0"
openai = "^1.75.0"
langchain = "^0.3.23"
langchain-openai = "^0.3.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
